Example, documentation, debug statements improved
EtherEvent. added to the start of debug statements to identify them from
this library.
Preprocessor directive for the w5100 timeout setup code. I need to test
if the include will cause problems with ENC28J60 use or if they should
comment that out because I can't use the preprocessor directive on
includes