Use a control character as the default password value
The goal with this is to allow the user to pass an empty password string
("") to the availableEvent or send functions to selectively disable
authentication. Previously I was using an empty string as the default
value. Because the use of the standard password set via setPassword() is
triggered when the default value is passed this prevented a blank
password string from being intentionally passed. The control character I
chose is "form feed", which is unlikely to ever be used as a password by
the user.