Document that Printable type is not supported by send() in FAST_SEND mode
Adding Printable type support would require me to have fixed global
event and payload buffers as there is no way for me to create a buffer
just for the Printable to char array conversion. I already support the
most common Printable type, IPAddress and Printable is supported in
non-FAST_SEND mode so I don't think this will be much of a limitation.