Send each payload wrapper as a single packet
In unauthenticated mode the payload is wrapped in ['']. The reason is
that TCPEvents assumes TCPEvents server type in unauthenticated mode and
will try to evaluate as Python any payload not wrapped like this.

Unfortunately this does increase global SRAM usage by 6 bytes due to the
string literals but it reduces flash memory usage (~24 bytes), makes the
send faster, and makes the code more readable. I don't want to use the
F() macro because that will cause each character to be sent in a
separate packet.