Only prepend TCPEvents identifier to accept message when necessary
TCPEvents prepends "TCPEvents" to the MD5 to identify itself. In
response I should prepend a space to the accept message to identify the
receiver as TCPEvents type (necessary in order to support TCPEvents
custom prefix option in the Send Event action configuration). Sending
this space to Network Event Sender plugin causes not problems but it
causes the wrong event to be received from the
timhoeck/eventghost-android app. The solution is to only send the space
when the received MD5 had "TCPEvents prepended.