Change to a header-only library
The use of templates forces much of the source code to be located in the .h file.

The use of preprocessor configuration from the sketch forces even more of the source code to be located in the .h file and makes it more difficult to use a .cpp file.

The scattering of source code results in the library being more difficult to maintain.

Changing to header-only library reduces memory usage.

The downside is possible increase of compilation time but that shouldn't be very significant.