Don't specify language property in Travis config
Travis CI has started their rollout of the Trusty Linux distribution. For some reason when that distribution is used and the language is specified as bash the xvfb virtual display configuration fails, causing the build to later fail when an attempt to run arduino is made. In this usage there really is not a valid language to set and not specifying any language solves the problem so I think this is the best solution.