Improve code that prevents overflow of availableEventSubmessageLengthMax
The previous code relied on unsigned overflow, which is undefined behavior. Instead I used the logical type size_t for availableEventSubmessageLengthMax and just use SIZE_MAX. I tried using the nice code (typeof(availableEventSubmessageLengthMax)-1 to determine the max size of availableEventSubmessageLengthMax, which I prefer because it will work for any type availableEventSubmessageLengthMax might be changed to but typeof() is GCC specific so this caused compilation for ARC32 to fail.