Add a HIL test library
Previously the test sketches were primarily used for compilation testing without having effective mechanisms to check whether the library actually works. This change adds "hardware in loop" (HIL) test functionality that will allow the tests to be run in conjunction with special EventGhost trees in order to verify that the sent data matches the received data. The code is put in a library to enable it to be shared between a sketch for each combination of EtherEvent library preprocessor configurations.