Wrap payload in send() conversion overloads when in fast send, unauthenticated mode
This allows faster sending of events because the wrapper strings don't need to be sent as separate packets.
This is more efficient because it allows me to only check for ' in the payload when it is of char array, char, __FlashStringHelper, or String type. I already know that character will never appear in a number so I can skip the check in those conversion functions. When the check was done in the main send() it had to be done for every payload since I couldn't know their original type.

The wrapping still needs to be done in the main send() if in authenticated mode but the password argument was "" (per-send unauthenticated).

The check for ' in payload can't be done in non-fast send mode because I don't have any way to know the type of payload.